This project is for building an AI Agent APP with Python, FastAPI, Langchain framework, Open  Source LLM (mistral) etc.
This project is for building an AI Agent APP with Python, FastAPI, Langchain framework, Open Source LLM (mistral) etc. As we know, knowledge of any LLM is restricted to its training data
and also till the date/time when it was trained and built. So, if we ask any question related to current/recent time (like tomorrow's weather) or current/recent affairs (like current job market in 2024)
to LLM, it fails to answer. An AI Agent solves this problem. Here, we create an ai agent and also empower it with external tools (like Wikipedia, serpapi (for search engine such as Google, Bing etc.)) along with LLM. Then we integrate this ai agent to a simple App created with FastAPI and it answers all of those queries satisfactorily. In this project, many features from Langchain are used. Open Source LLM "mistral" is locally used
through Ollama tool.